Refactor listeners. Do not push service manager and use listener aggreagte
